# openapi.yaml
openapi: 3.1.0
info:
  title: Lost&Found API
  version: 0.1.0
  description: Standalone web API. Public advertisement responses never include owner contacts.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/ads:
    get:
      operationId: listAds
      parameters:
        - { name: type, in: query, schema: { $ref: '#/components/schemas/AdType' } }
        - { name: category, in: query, schema: { type: string } }
        - { name: lat, in: query, schema: { type: number, minimum: -90, maximum: 90 } }
        - { name: lng, in: query, schema: { type: number, minimum: -180, maximum: 180 } }
        - { name: radius, in: query, schema: { type: integer, minimum: 1, maximum: 50000, default: 5000 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        '200':
          description: Active public advertisements
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdSummary' }
    post:
      operationId: createAd
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAdRequest' }
      responses:
        '201':
          description: Advertisement created
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties: { id: { type: string, format: uuid } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
...
  /api/v1/ads/{id}:
    get:
      operationId: getAd
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Public advertisement details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdDetail' }
        '404': { $ref: '#/components/responses/Error' }
    patch:
      operationId: updateAd
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, minLength: 1, maxLength: 255 }
                category: { type: string, maxLength: 64 }
                description: { type: string, maxLength: 3000 }
                occurredAt: { type: [string, 'null'], format: date-time }
                lat: { type: [number, 'null'], minimum: -90, maximum: 90 }
                lng: { type: [number, 'null'], minimum: -180, maximum: 180 }
      responses:
        '200': { description: Owned advertisement updated }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
  /api/v1/matches:
    get:
      operationId: listMatches
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses:
        '200':
          description: Ranked active matches involving an advertisement owned by the current user
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Match' } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /api/v1/matches/{id}/feedback:
    put:
      operationId: saveMatchFeedback
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [feedback]
              properties:
                feedback: { $ref: '#/components/schemas/MatchFeedback' }
      responses:
        '200': { description: Feedback stored as a separate dataset row }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/owner-checks:
    get:
      operationId: listOwnerChecks
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Claims where the current user is holder or claimant }
    post:
      operationId: submitOwnerCheck
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listingId, answers]
              properties:
                listingId: { type: string, format: uuid }
                answers:
                  type: array
                  minItems: 1
                  maxItems: 3
                  items:
                    type: object
                    required: [questionId, answer]
                    properties:
                      questionId: { type: string, format: uuid }
                      answer: { type: string, minLength: 1, maxLength: 2000 }
      responses:
        '201': { description: Owner check submitted }
        '409': { $ref: '#/components/responses/Error' }
  /api/v1/owner-checks/questions/{listingId}:
    get:
      operationId: getOwnerQuestions
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: listingId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Public verification questions without answers }
    put:
      operationId: replaceOwnerQuestions
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: listingId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompts]
              properties:
                prompts: { type: array, minItems: 1, maxItems: 3, items: { type: string, minLength: 1, maxLength: 500 } }
      responses:
        '200': { description: Questions replaced by the FOUND ad owner }
        '409': { $ref: '#/components/responses/Error' }
  /api/v1/owner-checks/{id}:
    get:
      operationId: getOwnerCheck
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Owner check, answer snapshots and allowed contact type names without contact values
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/owner-checks/{id}/decision:
    patch:
      operationId: decideOwnerCheck
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties: { decision: { type: string, enum: [APPROVED, DECLINED] } }
      responses:
        '200': { description: Pending check decided by the FOUND ad owner }
        '409': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/owner-checks/{id}/contacts:
    get:
      operationId: getOwnerCheckContacts
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Only contact values allowed by the FOUND listing owner, snapshotted at approval. A CONTACT_DISCLOSED audit event stores type names, never values.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DisclosedContacts' }
        '403': { $ref: '#/components/responses/Error' }
...
  /api/v1/reports:
    post:
      operationId: createReport
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listingId, reason]
              properties:
                listingId: { type: string, format: uuid }
                reason: { $ref: '#/components/schemas/ReportReason' }
                details: { type: string, maxLength: 2000 }
      responses:
        '201': { description: Report queued for moderation }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
...
  /api/v1/privacy-requests:
    post:
      operationId: createPrivacyRequest
      description: Creates a data-subject request. When authenticated, the account email is used and the submitted email is ignored.
      security: [{ bearerAuth: [] }, {}]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePrivacyRequest' }
      responses:
        '201':
          description: Request registered without echoing its sensitive content
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PrivacyRequestReceipt' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
  /api/v1/admin/reports:
    get:
      operationId: listAdminReports
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/ReportStatus' } }
      responses:
        '200': { description: Moderation report queue }
        '403': { $ref: '#/components/responses/Error' }
  /api/v1/admin/reports/{id}:
    patch:
      operationId: updateAdminReport
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [IN_REVIEW, RESOLVED, REJECTED] }
                resolutionNote: { type: string, maxLength: 2000 }
      responses:
        '200': { description: Report status changed and audited }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /api/v1/admin/users:
    get:
      operationId: listAdminUsers
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Administrative user list }
        '403': { $ref: '#/components/responses/Error' }
  /api/v1/admin/users/{id}:
    patch:
      operationId: moderateUser
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties: { action: { type: string, enum: [BLOCK, UNBLOCK] } }
      responses:
        '200': { description: User state changed and audited }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /api/v1/admin/ads:
    get:
      operationId: listAdminAds
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Administrative advertisement list }
        '403': { $ref: '#/components/responses/Error' }
  /api/v1/admin/ads/{id}:
    patch:
      operationId: moderateAd
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties: { action: { type: string, enum: [HIDE, RESTORE, DELETE] } }
      responses:
        '200': { description: Advertisement state changed and audited }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/admin/privacy-requests:
    get:
      operationId: listAdminPrivacyRequests
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/PrivacyRequestStatus' } }
      responses:
        '200':
          description: Data-subject request queue, restricted to administrators
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PrivacyRequest' }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /api/v1/admin/privacy-requests/{id}:
    patch:
      operationId: updateAdminPrivacyRequest
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdatePrivacyRequest' }
      responses:
        '200': { description: Request status changed and action audited without request content }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Error:
      description: Request error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    AdType: { type: string, enum: [LOST, FOUND] }
    MatchFeedback: { type: string, enum: [CONFIRMED, SIMILAR_NOT_MINE, NOT_RELEVANT] }
    ReportReason: { type: string, enum: [SPAM, FRAUD, INAPPROPRIATE, DUPLICATE, OTHER] }
    ReportStatus: { type: string, enum: [OPEN, IN_REVIEW, RESOLVED, REJECTED] }
    PrivacyRequestType: { type: string, enum: [ACCESS, CORRECTION, DELETION, WITHDRAW_CONSENT, OTHER] }
    PrivacyRequestStatus: { type: string, enum: [NEW, IN_PROGRESS, RESOLVED, REJECTED] }
    CreatePrivacyRequest:
      type: object
      required: [requestType, message]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
          description: Required without authentication; ignored when a bearer token identifies an active account.
        requestType: { $ref: '#/components/schemas/PrivacyRequestType' }
        message: { type: string, minLength: 10, maxLength: 3000 }
    PrivacyRequestReceipt:
      type: object
      required: [id, status]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, const: NEW }
    PrivacyRequest:
      type: object
      required: [id, requestType, message, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        userId: { type: [string, 'null'], format: uuid }
        email: { type: [string, 'null'], format: email }
        requestType: { $ref: '#/components/schemas/PrivacyRequestType' }
        message: { type: string }
        status: { $ref: '#/components/schemas/PrivacyRequestStatus' }
        resolutionNote: { type: [string, 'null'] }
        resolvedBy: { type: [string, 'null'], format: uuid }
        resolvedAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    UpdatePrivacyRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [IN_PROGRESS, RESOLVED, REJECTED] }
        resolutionNote: { type: string, maxLength: 2000 }
    Match:
      type: object
      required: [id, score, algorithmVersion, lost, found]
      properties:
        id: { type: string, format: uuid }
        score: { type: integer, minimum: 0, maximum: 100 }
        algorithmVersion: { type: string, example: baseline-v1 }
        scoreBreakdown: { type: [object, 'null'], additionalProperties: true }
        feedback: { anyOf: [{ $ref: '#/components/schemas/MatchFeedback' }, { type: 'null' }] }
        ownsLost: { type: boolean }
        ownsFound: { type: boolean }
        lost: { $ref: '#/components/schemas/MatchAd' }
        found: { $ref: '#/components/schemas/MatchAd' }
    MatchAd:
      type: object
      required: [id, title, category]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        category: { type: string }
        occurredAt: { type: [string, 'null'], format: date-time }
        photo: { type: [string, 'null'] }
...
    CreateAdRequest:
      type: object
      required: [type, category, title]
      properties:
        type: { $ref: '#/components/schemas/AdType' }
        category: { type: string }
        title: { type: string, minLength: 1, maxLength: 255 }
        description: { type: string, maxLength: 3000 }
        lat: { type: [number, 'null'], minimum: -90, maximum: 90 }
        lng: { type: [number, 'null'], minimum: -180, maximum: 180 }
        occurredAt: { type: [string, 'null'], format: date-time }
        contactChannels:
          type: array
          uniqueItems: true
          minItems: 1
          maxItems: 3
          items: { $ref: '#/components/schemas/ContactChannel' }
          description: Required for FOUND listings; ignored only when empty for LOST listings. Values themselves never appear in public ad DTOs.
    ContactChannel:
      type: string
      enum: [EMAIL, PHONE, TELEGRAM]
    DisclosedContacts:
      type: object
      properties:
        email: { type: [string, 'null'], format: email }
        phone: { type: [string, 'null'] }
        telegram: { type: [string, 'null'] }
      description: Non-selected channels are null and contact values are available only to approved owner-check participants.
    AdSummary:
      type: object
      required: [id, type, category, title, created_at]
      properties:
        id: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/AdType' }
        category: { type: string }
        title: { type: string }
        description: { type: string }
        lat: { type: [number, 'null'] }
        lng: { type: [number, 'null'] }
        occurred_at: { type: [string, 'null'], format: date-time }
        created_at: { type: string, format: date-time }
        preview_photo: { type: [string, 'null'] }
    AdDetail:
      allOf:
        - { $ref: '#/components/schemas/AdSummary' }
        - type: object
          properties:
            status: { type: string, enum: [ACTIVE] }
            district: { type: [string, 'null'] }
            updated_at: { type: [string, 'null'], format: date-time }
            photos: { type: array, items: { type: string } }
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        message: { type: string }